# Add `selected_message_queues` flag to filter message queues and improve performances