--- a/ibm_i/assets/configuration/spec.yaml
+++ b/ibm_i/assets/configuration/spec.yaml
@@ -94,14 +94,21 @@ files:
           - name: "job_queue"
           - name: "message_queue_info"
         default: *queries_default
-    - name: selected_message_queues
+    - name: message_queue_info
       description: |
-        Filters the message queues to make your queries less computer-intensive. By default no filter is applied.
+        A configuration mapping for the message_queue_info query.
+
+        selected_message_queues filters the message queues to make your queries less computer-intensive.
+        By default no filter is applied.
       value:
-        type: array
-        items:
-          type: string
         example:
-          - QSYSOPR
-          - QPGMR
+          selected_message_queues:
+            - QSYSOPR
+            - QPGMR
+        type: object
+        properties:
+          - name: selected_message_queues
+            type: array
+            items:
+              type: string
     - template: instances/default