--- a/ibm_i/assets/configuration/spec.yaml
+++ b/ibm_i/assets/configuration/spec.yaml
@@ -98,8 +98,8 @@ files:
       description: |
         A configuration mapping for the message_queue_info query.
 
-        selected_message_queues filters the message queues to make your queries less computer-intensive.
-        By default no filter is applied.
+        selected_message_queues restricts the message_queue_info query to a subset of the queues.
+        This can reduce the CPU usage on the target IBM i machine. By default no filter is applied.
       value:
         example:
           selected_message_queues: