--- a/ibm_i/datadog_checks/ibm_i/config_models/defaults.py
+++ b/ibm_i/datadog_checks/ibm_i/config_models/defaults.py
@@ -38,6 +38,10 @@ def instance_job_query_timeout(field, value):
     return 240
 
 
+def instance_message_queue_info(field, value):
+    return get_default_field_value(field, value)
+
+
 def instance_metric_patterns(field, value):
     return get_default_field_value(field, value)
 
@@ -66,10 +70,6 @@ def instance_queries(field, value):
 
 def instance_query_timeout(field, value):
     return 30
-
-
-def instance_selected_message_queues(field, value):
-    return get_default_field_value(field, value)
 
 
 def instance_service(field, value):