--- a/ibm_i/datadog_checks/ibm_i/data/conf.yaml.example
+++ b/ibm_i/datadog_checks/ibm_i/data/conf.yaml.example
@@ -78,6 +78,13 @@ instances:
     #   - name: job_queue
     #   - name: message_queue_info
 
+    ## @param selected_message_queues - list of strings - optional
+    ## Filters the message queues to make your queries less computer-intensive. By default no filter is applied.
+    #
+    # selected_message_queues:
+    #   - QSYSOPR
+    #   - QPGMR
+
     ## @param tags - list of strings - optional
     ## A list of tags to attach to every metric and service check emitted by this instance.
     ##
@@ -86,13 +93,6 @@ instances:
     # tags:
     #   - <KEY_1>:<VALUE_1>
     #   - <KEY_2>:<VALUE_2>
-
-    ## @param selected_message_queues - list of strings - optional
-    ## Filters the message queues you're watching to make your queries less computer-intensive. By default no filter is applied.
-    #
-    # selected_message_queues:
-    #   - <MESSAGE_QUEUE_NAME_1>
-    #   - <MESSAGE_QUEUE_NAME_2>
 
     ## @param service - string - optional
     ## Attach the tag `service:<SERVICE>` to every metric, event, and service check emitted by this integration.