--- a/ibm_i/datadog_checks/ibm_i/queries.py
+++ b/ibm_i/datadog_checks/ibm_i/queries.py
@@ -259,7 +259,9 @@ def get_job_queue_info(timeout):
 
 def get_message_queue_info(timeout, sev, selected_message_queues):
     message_queues_list = [f"'{elt}'" for elt in selected_message_queues]
-    message_queues_filter = f"WHERE MESSAGE_QUEUE_NAME IN ({', '.join(message_queues_list)})" if selected_message_queues else ""  # noqa:E501
+    message_queues_filter = (
+        f"WHERE MESSAGE_QUEUE_NAME IN ({', '.join(message_queues_list)})" if selected_message_queues else ""
+    )
     return {
         'name': 'message_queue_info',
         'query': {
@@ -289,5 +291,7 @@ def query_map(config: InstanceConfig):
         "job_memory_usage": get_job_memory_usage(config.job_query_timeout),
         "memory_info": get_memory_info(config.query_timeout),
         "job_queue": get_job_queue_info(config.query_timeout),
-        "message_queue_info": get_message_queue_info(config.system_mq_query_timeout, config.severity_threshold, config.selected_message_queues),  # noqa:E501
+        "message_queue_info": get_message_queue_info(
+            config.system_mq_query_timeout, config.severity_threshold, config.selected_message_queues
+        ),
     }