--- a/ibm_i/tests/test_sql_queries.py
+++ b/ibm_i/tests/test_sql_queries.py
@@ -16,19 +16,27 @@ def test_get_message_queue_info():
         message_queue_info=MessageQueueInfo(selected_message_queues=[])
     )
     qmap_output = query_map(instance_conf)
-    assert qmap_output["message_queue_info"] == (
+    assert qmap_output['message_queue_info']['name'] == 'message_queue_info'
+    assert qmap_output['message_queue_info']['columns'] == [
+        {'name': 'message_queue_name', 'type': 'tag'},
+        {'name': 'message_queue_library', 'type': 'tag'},
+        {'name': 'ibm_i.message_queue.size', 'type': 'gauge'},
+        {'name': 'ibm_i.message_queue.critical_size', 'type': 'gauge'},
+    ]
+    assert qmap_output['message_queue_info']['query']['text'] == (
         'SELECT MESSAGE_QUEUE_NAME, MESSAGE_QUEUE_LIBRARY, COUNT(*), SUM(CASE WHEN SEVERITY >= 50 THEN 1 ELSE 0 END) '  # noqa:E501
         'FROM QSYS2.MESSAGE_QUEUE_INFO GROUP BY MESSAGE_QUEUE_NAME, MESSAGE_QUEUE_LIBRARY'
     )
+    assert qmap_output['message_queue_info']['query']['text'] == 1
     instance_conf.message_queue_info = MessageQueueInfo(selected_message_queues=['QSYSOPR'])
     qmap_output = query_map(instance_conf)
-    assert qmap_output["message_queue_info"] == (
+    assert qmap_output['message_queue_info']['query']['text'] == (
         'SELECT MESSAGE_QUEUE_NAME, MESSAGE_QUEUE_LIBRARY, COUNT(*), SUM(CASE WHEN SEVERITY >= 20 THEN 1 ELSE 0 END) '
         'FROM QSYS2.MESSAGE_QUEUE_INFO WHERE MESSAGE_QUEUE_NAME IN (\'QSYSOPR\') GROUP BY MESSAGE_QUEUE_NAME, MESSAGE_QUEUE_LIBRARY'  # noqa:E501
     )
     instance_conf.message_queue_info = MessageQueueInfo(selected_message_queues=['QSYSOPR', 'QPGMR', 'CECUSER'])
     qmap_output = query_map(instance_conf)
-    assert qmap_output["message_queue_info"] == (
+    assert qmap_output['message_queue_info']['query']['text'] == (
         'SELECT MESSAGE_QUEUE_NAME, MESSAGE_QUEUE_LIBRARY, COUNT(*), SUM(CASE WHEN SEVERITY >= 30 THEN 1 ELSE 0 END) '
         'FROM QSYS2.MESSAGE_QUEUE_INFO WHERE MESSAGE_QUEUE_NAME IN (\'QSYSOPR\', \'QPGMR\', \'CECUSER\') GROUP BY MESSAGE_QUEUE_NAME, MESSAGE_QUEUE_LIBRARY'  # noqa:E501
     )